requirements.txt
python-dotenv
PyGithub
rich
urllib3==1.26.16
fastapi
uvicorn[standard]
